requirements.txt
asgiref==3.7.2
boto3==1.34.19
botocore==1.34.19
dj-database-url==2.1.0
Django==5.0
django-multiselectfield==0.1.12
django-storages==1.14.2
djangorestframework==3.14.0
gunicorn==21.2.0
jmespath==1.0.1
packaging==23.2
pillow==10.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7
whitenoise==6.6.0
